refactor and introduced utils/locale.js to get/set current locale
