+++ frontend/src/utils/documentListHelper.js
@@ -2,11 +2,11 @@ import PropTypes from 'prop-types';
 import { Map as iMap } from 'immutable';
 import Moment from 'moment-timezone';
 import currentDevice from 'current-device';
-import { LOCAL_LANG } from '../constants/Constants';
 import { getItemsByProperty, nullToEmptyStrings } from './index';
 import { getSelectionInstant } from '../reducers/windowHandler';
 import { viewState, getView } from '../reducers/viewHandler';
 import { TIME_REGEX_TEST } from '../constants/Constants';
+import { getCurrentActiveLocale } from './locale';
 
 /**
  * @typedef {object} Props Component props
@@ -318,14 +318,6 @@ export function getScope(isModal) {
 
 export function parseToDisplay(fieldsByName) {
   return parseDateToReadable(nullToEmptyStrings(fieldsByName));
-}
-
-/**
- * @method getCurrentActiveLocale
- * @summary Retrieves the active locale from the local store
- */
-export function getCurrentActiveLocale() {
-  return localStorage.getItem(LOCAL_LANG);
 }
 
 export function convertTimeStringToMoment(value) {
